+++ exercicio2/src/main/java/br/com/andre/exercicio2/business/Result.java
@@ -5,11 +5,11 @@ import java.util.List;
 import lombok.Getter;
 
 public class Result {
-    private final @Getter int number;
+    private final @Getter long number;
     private final @Getter boolean isHappy;
     private final @Getter List<Step> steps;
 
-	public Result(int number, boolean isHappy, List<Step> steps) {
+	public Result(long number, boolean isHappy, List<Step> steps) {
         this.number = number;
         this.isHappy = isHappy;
         this.steps = steps;
